ser-experiment/requirements.txt
torch==1.9.0+cu111
numpy==1.19.5
torchvision==0.10.0+cu111
tqdm==4.62.3
pandas==1.3.2
Pillow==7.0.0